fix the wrong websocket URL (#920)
Previously the WebSocket URL depends on the window.location.href. It is the current URL in the browser. When the workspace is requested by some other URLs, e.g., /workflow/:id, it is no longer relevant to use this as part of WebSocket URL.

Thus this changes window.location.href to window.location.origin to temporarily fix the issue.