Data Processing Module Refactoring (#900)
### **Overall:**
1. removed old data processing functionality from Generator and Processor class.
2. created a `neo` folder under `worker` folder and put all refactored data processing functionality inside it.
3. adapted new data processing functionality with the current worker design.

### **What's changed in Amber:**

Now the worker class has 5 components:
1. **WorkerInternalQueue Component**:  receive batches in FIFO order from the network layer.
2. **BatchToTupleConverter Component**:  read batches and output tuples into DataProcessing Module.
3. **DataProcessing Component**: contain DP thread, process tuples from BatchToTupleConverter and output tuples to TupleToBatchConverter.
4. **PauseControl Component**: provides pause and resume functionality for both actor and DP thread.
5. **TupleToBatchConverter Component**: for now, it's the old DataTransferSupport. I don't want to make this PR too big.

### **Decisions I've made:**
 
This PR is the first step of the whole refactoring process. I was trying to do "surgery" on the engine instead of directly modifying the project structure. 

~I utilized the "[cake pattern](https://www.baeldung.com/scala/cake-pattern)" in scala to keep everything modulized. I encourage everyone to use it in this round of refactoring~

There are a few articles attack cake pattern: [The cake pattern is a lie](https://www.mikulskibartosz.name/the-cake-pattern-is-a-lie/), [Cake antipattern](https://kubuszok.com/2018/cake-antipattern/). It turns out that the cake pattern is not a good solution for dependency injection.

The downsides of cake pattern: 
1. Since every component is a `trait`, which means it cannot be instantiated. So it's hard for developers to write test cases for each component.
2. The cake pattern mixes all the components together, which means there will be no more "layered" structure. The "mixed" class will be too huge to maintain. i.e. The "cake size" will grow bigger and bigger when we have more components. 

Thus, I decided to use [MacWire](https://github.com/softwaremill/macwire), which is a mature dependency injection framework in scala under Apache2.0 license. it scans the current scope in compile-time and injects dependencies for the developers. If some are missing, compilation fails. Developers meanwhile just write `wire[ClassName]`. It's very easy to use and convenient.

Co-authored-by: Shengquan Ni <Hiseen@github.noreply.com>