Added operator grouping feature (#754)
Added operator grouping feature
https://github.com/Texera/texera/pull/754

Authored-by: Abigail (Yingyu) Yang <yangyingyu10@gmail.com>
Co-authored-by: Albert Liu <ayliu5@uci.edu>
Co-authored-by: Zuozhi Wang<zuozhiw@gmail.com>

## **Couple rules that restrict grouping:**
1. In-links of a group are links whose source operators are outside of the group, and target operators are inside of the group.
2. Out-links of a group are links whose target operators are outside of the group, and source operators are inside of the group. 
3. There are always at least two operators in a group.
4. Groups can't be overlapping: no group can be created inside of another group.
5. When a group is expanded, user can delete operators in the group, and add & delete links, in-links & out-links of the group.
6. When a group is collapsed, user can only delete in-links and out-links of the group, and add those links back with undo (undo is not yet implemented).
7. Once a group is created, user can't add new operators to the group (plan to change this later)


## **Group operators:**
User can group multiple operators together by selecting them and clicking on the group button in nav bar. All selected operators are considered as part of the group. Links whose source and target operators are both in the group are also considered as part of the group. A newly added group and its in-links & out-links are placed in front of everything else on the graph.

## **Ungroup operators:**
When there are less than two operators in a group, the group will be automatically ungrouped and remaining operator and links will be freed.

## **Collapse group:**
User can collapse a group by clicking on the minus sign on the top-right corner of a group. When a group is collapsed, all operators and links in the group are hidden (temporarily deleted from joint graph, but still exist in texera graph). The group's in-links & out-links will be redirected to connect to the group element instead.

## **Expand group:**
User can expand a group by clicking on the plus sign button on the top-right corner of a collapsed group. When a group is expanded, operators and links that are hidden will be re-shown on the graph. In-links & out-links will be directed back to their original operators.

## **Moving group & operators:**
When an operator inside a group moves, the group will be repositioned and resized based on where the operator is moved to. So the group will follow its operators as they move. When a group is moved, all operators inside it moves as well, regardless of whether is expanded or collapsed.

## **Delete operators from group:**
User can only delete an operator embedded in a group when the group is expanded. When the operator is deleted, the group will be resized to fit remaining operators.

## **Add & delete links/in-links/out-links in groups:**
User can only add & delete a link, in-link, or out-link of a group when the group is expanded. When the group is collapsed, user can only delete in-links and out-links.

## **Groups in drag-drop:**
The drag drop service will ignore operators that are in collapsed groups. So when a new operator is dragged to the graph, it will not try to make suggestions for operators that are not on the joint graph.

## **Groups in delete all feature:**
When user tries to delete all operators/elements on the graph using the delete all button on nav bar, operators and links in both collapsed groups and expanded groups will be deleted as well.

## **Groups in save-workflow:**
Groups are saved to the local storage exactly as how they last appeared on the graph. When the browser is refreshed, expanded groups are still expanded, and collapsed groups are still collapsed.