Improves multiple ports support in Texera operators (Step 4)  (#982)
This PR changes the execution of a workflow to make the execution be aware of the ports. 
Previously in step 1, where there are no concept of ports in the backend, the mapping between an input operator to its corresponding ordinal is decided by the order that the connection is made. This PR changes it to be the actual port that the operator is connected to.