Let PythonUDF take empty input tuples (#1087)
This PR fixes an issue when receiving no tuple from upstream, it will break. The schema is obtained during runtime, so without tuple provided in runtime, python UDF is not aware of the input schema, thus cannot convert a null schema to Arrow Schema. 

The solution now is just do not send out any flight (which requires Arrow Schema) if no tuples are to be sent.

<sup> Created from JetBrains using [CodeStream](https://codestream.com/?utm_source=cs&utm_medium=pr&utm_campaign=github*com)</sup>