Progressive Result Display  (#1139)
This task is done by Tongjie Wang . Link to original PR: #1007. Pull request reopened because the branch is rebased on master.
Zuozhi reviewed this PR and approved it. Tongjie further verified the changes and approved it.

The server will send status updates (with the newest 10 rows of output) to the client.

The client will update the result panel once per second based on the received status update.

![Screen Recording 2021-01-31 at 11 10 33 PM](https://user-images.githubusercontent.com/21030118/106427020-78e37800-641b-11eb-9350-88dbbcba3cf0.gif)

It can also display not only appended result, but also modified and deleted in real time.

https://user-images.githubusercontent.com/21030118/119858066-b332d100-bf46-11eb-98da-c53cad58a6c8.mp4

Authored-by: Tongjie <tjwang0221@gmail.com>