Minimap relocation (#1185)
According to user feedback, the minimap is too large and obstruct the main workflow canvas.

This PR makes the minimap smaller and moves it to back to the bottom right corner, below the property editor. The property editor is also now wider. The new UI looks like this:

<img width="1606" alt="Screen Shot 2021-06-25 at 23 02 37" src="https://user-images.githubusercontent.com/17627829/123503706-77545f80-d609-11eb-8918-92176e605b82.png">


Old UI for comparison:
<img width="1606" alt="Screen Shot 2021-06-25 at 23 02 17" src="https://user-images.githubusercontent.com/17627829/123503712-7b807d00-d609-11eb-9dcb-9b2cc050790d.png">



Co-authored-by: Yicong Huang <17627829+Yicong-Huang@users.noreply.github.com>