Disable WorkflowCache when UserSystem is enabled (#1288)
The previous implementation has `WorkflowCache` and `WorkflowPersist` services enabled in parallel. This creates unsynced issues when the network is slow and persisted workflow is not synced in the cache. 

This PR removes this parallelism: 
- when `UserSystem` is enabled, normally in production mode, the `WorkflowPersistService` will be enabled.
- when `UserSystem` is disabled, normally in development mode, the `WorkflowCacheService` will be enabled.

The `WorkflowPersistService` and the `WorkflowCacheService` become supplementary to each other and will never interact with each other.

There are some more changes in this PR:
- Simplifies the logic of creating a new workflow, combining `/workflow/new` onto `/` to indicate creating of a new workflow.
- Disables the workspace while loading a workflow.