Limit console message displayed (#1298)
When too many console messages are shown, it could become a memory burden for the frontend. This PR solves the memory burden by limiting the number of console messages that can be displayed for each operator (for now, only Python UDF V2). The implementation uses a circular buffer to show the latest 100 messages for each operator.