Allow print in Python UDF V2 open() (#1421)
Previously, the operator is opened during the initialization of the worker and its operator. This is unclear semantic especially we need to much more things during `open()`, such as print, and load a large model. 

This PR introduces a new command `OpenOperator` to be sent after the worker and operator initialization. Fixes #1401.

right now the initializing logic follows: 
```
InitializeOperatorLogic() -> LinkWorkers() -> actor running -> OpenOperator() -> WorkflowReady
```