Allow users to rename projected columns (#1459)
Users now can rename the columns they want to project on. Click on the + button and two input boxes will appear.  The first box is a drop-down menu that contains all of the names of the attributes in the schema. This box is required. The second box is an input box that allows the user to type in the alias for the attribute they select. This box is optional. 


![projectionRename](https://user-images.githubusercontent.com/57199722/155591803-88cd7c8c-da2b-4581-996b-9626ec5d58bb.gif)

Co-authored-by: sadeemsaleh <sadeem.alsudais@gmail.com>