Temp fix for operator cyclic dependency (#1504)
There are some operators have dependency on input ports (HashJoin, 2-in Python UDF, etc). When there are two instances of the operator, they can introduce more than one dependencies and there could exist cycles.

For example, in the following workflow, we have a "left-only-plan" on using models onto tuples.
<img width="538" alt="Screen Shot 2022-04-23 at 8 44 21 AM" src="https://user-images.githubusercontent.com/17627829/164913180-6d5b49f7-3b12-4290-86db-c7d7a700a6c4.png">

C needs A to run before B starts (B depends on A). 
D needs A to run before C starts (C depends on A).

The current implementation when C depends on A, it will find all sources of the depender C (A and B in this case) and introduce wait on all sources to the dependee, thus A will wait for A, introducing a cycle.

This PR fixes it by ignoring "self-waiting" cycle.

There are more cyclic dependency cases to be solved (e.g., #1324). This PR only covers the above simple case, to make sure the left-join-only plan can execute properly.