Adding credit concept to worker (#1516)
This is the 1st PR to bring flow control in Texera. This PR defines the concept of credits in a worker for each sender worker. Specifically,

Initial fixed credit amount is given to each sender worker
When a data item from a sender is put into internal queue, the credit is decreased by 1.
When a data item is taken out of the internal queue to be processed, its credit is increased by 1.