Allow multiple users to share workflow execution (#1635)
This PR changed the workflow service to make multiple users share the same workflow executions. 
In the existing implementation, user A and user B can have two different executions of the same workflow at the same time. 
In the new implementation, both users see the same execution of the same workflow at the same time.