Introduce engine version for workflow executions (#1617)
- [x] fix issue that the file doesn't exist when running the APP. either manually create it and add it to the PR, or write a piece of code to handle creating the file for the first time

This PR introduces Texera engine version to each workflow execution. The engine version is a hash code created automatically when the Texera is run. The engine version is not shown in the front end but can be checked in the sql for now.
The database schema has changed -- a new column "environment_version" is added to the table workflow_executions. The corresponding ALTER TABLE command is: `ALTER TABLE workflow_executions ADD COLUMN environment_version VARCHAR(128) NOT NULL`.

- [x] need to test if the version is changed, PR #1503 will highlight the label.

Co-authored-by: Liqin Ye <liqiny@uci.edu>
Co-authored-by: sadeem <sadeem.alsudais@gmail.com>