Fix regex operator serialization issue (#1689)
The regex operator executor holds a `Pattern` object. Currently the executor instance is created at the controller site, and then shipped to the worker side. However the `Pattern` object cannot be propertly serailized (case insensitive match flag is lost). 

This PR fixes this problem by changing from `val` to `lazy val`, so the `Pattern` instance is created at the worker side (when the worker processes the first input tuple).