Fix NaN in Tuple and Table by treating them as None (#1703)
In pandas, numpy, the null(missing) values in numerical types are treated as `NaN`. There are some usages regarding `NaN` in the context of pandas and numpy, thus it is a good idea to keep them within Tuple and Table within the PVM. However, it is troublesome to transfer `NaN` through Arrow: 
1. if `NaN` present, the dtype of the column becomes float. Arrow is not able to force cast a float `NaN` to an integral type.
2. `NaN` is not comparable, which means `NaN == NaN` will return false, while `None == None` would return true, which can cause issues in value check and conversions.

This PR leaves `NaN` within the Python context, but before transferring through Arrow back to java side, it converts all `NaN` values to `None`. On Java side, those values will be `null` and preserve the integral types if applicable. 

The issue was seen via https://github.com/Texera/texera/pull/1690.

Co-authored-by: Xiaozhen Liu <xiaozl3@uci.edu>