Amber Fault Tolerance: Logging (#1554)
This PR includes a full lifecycle of log-based fault tolerance:

Added LogStorage, which is an abstraction of where the log is persisted. Currently implemented LocalFS and HDFS.
Added AsyncLogWriter, which is a thread that pushes the serialized log records into LogStorage.
Added SerializationManager, which does the compression and conversion of determinant from in-mem format to byte array, controlled by the AsyncLogWriter.
Added LogManager, which handles the logging for both data and control determinants and manages the lifecycle of AsyncLogWriter and LogStorage.
Added TimeService and OperatorContext, the operator can access the time service through the context in its own logic to get the timestamp logged by the log manager. We can also use the same context to put the pause manager if we want to let the operator pause itself. Right now due to some scala-java incompatibility issues of trait, I cannot put a member variable inside IOperatorExecutor. We need to do code refactoring for this.