Add autocomplete to filename in file scan operators (#1688)
![Peek 2022-10-17 22-23](https://user-images.githubusercontent.com/114192306/196342621-d8f19308-812a-4e42-8737-cb04b2a1c7ed.gif)

This PR contains the function of autocomplete input box. which is mainly for file input. It's a custom Formly component that can also be implemented in other file input places. Users can type username, file name, or username plus filename. The filenames that contain the input content have the highest priority, followed by the username and username plus filename. 

Currently, the file list is handled by the backend. The current usage is a keyword search. In order to process the file path symbol "/" and reserved symbols such as " ", the frontend will encode the input and backend will decode the input, which can process all kinds of input.

Co-authored-by: Yicong Huang <17627829+Yicong-Huang@users.noreply.github.com>