Introduce Debug Command to Worker (#1754)
This PR adds `DebugCommand`, a control message that delivers debugging-related instructions to the debugger. 

In the first step, we will handle `DebugCommand` only on Python UDF workers. The same API can be used to debug native operators in the future. The backend logic of handling such `DebugCommand` is to be added when integrating a debugger (e.g., pdb), which will be covered in later PRs.

## DebugCommand
The current implementation expects the command to be a raw string. It is general to accept any format that the backend debugger requires. A Debug Command will result in a ConsoleMessage, with the type as `COMMAND` and the source to be `USER-{uid}`.

## Command Input Box & Worker Selection Menu
Added an input box to the Console of UI. Users can type a command in the input box and send it anytime during the execution, regardless of the workflow or operator state. The input command will be displayed on the Console right after the user inputs it, as a `ConsoleMessage`. Besides the input box, there is a dropdown menu to select the target worker (or all workers) to send to.

## Expose WorkerId to frontend
As debugging is per worker level, this PR is the first time we formally expose WorkerId to frontend so that the user can choose which worker to debug on. This is mainly for experienced users only. 

## Demo
<img width="656" alt="Screenshot 2022-12-01 at 22 26 50" src="https://user-images.githubusercontent.com/17627829/205232125-084b6947-aa3f-4818-9478-0023b5ba0817.png">