Add Descriptions to Projects (#1767)
Introduction 
---
This PR takes steps to improve the project feature by adding an editable
description. The intent is that project descriptions will be more
complex and comprehensive than descriptions of workflows or files.

Solution
---
The existing description feature on workflows and files was the
inspiration behind this feature.
The following changes were made to implement this description feature

- SQL DDL change to add `description` column to `user_project` table
- backend web application change to `ProjectResource.scala` to add
endpoint for updating descriptions
- frontend changes to render editable project description, modeled off
of Ng Zorro Typography component

Demo
---

https://user-images.githubusercontent.com/57075492/206831905-76e1b8e5-d404-49a6-bae3-b4f491f44fc6.mov


Current Issues
---
This is an issue with the existing workflow / file description feature
as well, but text does not wrap when it gets too long and the
description component currently does not automatically focus (users have
to click into the text box to start typing, the cursor is not
automatically moved there for them). These are all edge cases but fixing
them could improve the user experience.

Future TODOs
---
- Add capabilities to edit project description from within a single
`UserProjectSection` component as well
- Make lengthy user descriptions wrap instead of overflowing
- Potentially improve on description rendering by adding support for
markup
- Add some interval to potentially save descriptions at different
checkpoints as Yicong suggested