Fix input port mapping not passed into data processor in some cases (#1812)
This PR fixes an issue introduced in #1793 . In some scenarios, the
`inputPortMapping` is not correctly passed into the data processor due
to an initilaization order issue of a lazily evaluated variable. This
causes the join operator to not produce results because it wrongly
thinks all the input data are from the build side.

This PR does a temporary hot-fix of this issue, this issue will be
completely solved once #1807 is merged into master.