Use memory storage for HTML visualization sink operators (#1875)
The visualization sink usually accepts an HTML that could be large in
terms of size. If we use MongoDB as the sink storage, we cannot
visualize the HTML since the record cannot be inserted into MongoDB due
to a hard limit on the size of a single document (16MB). To fix this
issue, HTML visualization sinks will always use memory storage after
this PR as a quick solution.

In the future, we can break the HTML into multiple records and save them
into MongoDB, then merge them back when we do the visualization.