Broadcast Partition method for python and scala (#1891)
In this PR, we implemented a new partition method, which is broadcast.
Unlike to other method, this one will send the data to all the workers
so that they can do the task at the same time. A good example for
broadcast partition is 2-in-pythonUDF. When we have a model and want to
do prediction with multiple workers, we can use broadcast partition
method.

In order to do this, we changed the single int index to an iterator in
the partitioner, which can help us to return multiple indices (to
multiple workers). Also, we implemented broadcast method to both python
and scala parts. By local test on 200k data workflow, the iterator is
even a little bit faster (14 vs 17 seconds).

Here is a demo for broadcast partition in 2-in-pythonUDF. To simply the
process, we put "clf"="test" and "vectorize"=2 in the model port, and
let the prediction part print these value. As we can see, every worker
receives these value correctly.
![Peek 2023-04-24
20-59](https://user-images.githubusercontent.com/114192306/234171430-8d675ff5-8032-4eb6-9e48-8cd25a85727f.gif)