Refactored each workflow list item into its own component (#1937)
Refactored each workflow list item into its own component in preparation
for universal search. Related to #1935

Previously, all the logics for what you can see in the screenshot was in
`UserWorkflowComponent`. After the refactoring, the red box portion is
moved to `UserWorkflowListItemComponent`.
In the universal search UI, the new `UserWorkflowListItemComponent` will
be re-used.


![image](https://github.com/Texera/texera/assets/9412032/f24c0088-eecb-438c-a6f5-a89a72f67ced)