Enforce Node LTS v18 (#1942)
As Node 14.x has reached the end of life on
[04/30/2023](https://nodejs.dev/en/about/releases/), we need to migrate
to a newer LTS version. Node 16.x is currently under maintenance and has
also announced the early end of life on
[09/11/2023](https://nodejs.org/en/blog/announcements/nodejs16-eol),
which is in less than 4 months; it may be better to skip version 16.x.

Thus, we will start to enforce Node 18.x.