Validate UDF output tuple against the schema (#1952)
This PR validates the output Tuple's field values against the required
output Schema and reports schema mismatch issues.
1. checks tuple contains fields described in the schema;
2. checks if the value of each field matches that described in the
schema.

## Demo
<img width="1257" alt="Screenshot 2023-06-01 at 23 25 00"
src="https://github.com/Texera/texera/assets/17627829/411b247e-fefa-480d-bc3f-8b68b0172a48">


## Implementation

1. It validates schema after the creation of an output Tuple.
3. It currently considers Schema as an external value to check instead
of making Schema a property of the Tuple. This is to accommodate the
lazy Tuple design, which does not require a Tuple to have a fixed
Schema, and the values are only parsed upon usage for an input Tuple.
4. It also uses the `pyarrow.Schema` for now. We will consider wrapping
it with the Texera Schema definition and maintaining the mapping with
`pyarrow.Schema`.