Introducing Enhanced Version View with Difference Editor (#1991)
<b>Description</b>:
This PR introduces a new feature to Texera - an enhanced version view
with a robust Difference Editor. With this update, users can now
effortlessly visualize and comprehend the changes made between different
versions of our workflows.

<b>Key Features:</b>
1. Side-by-side Comparison: A seamless experience by viewing two
versions simultaneously, with the original on one side and the modified
version on the other.
2. Highlighting Changes: Clearly identify additions, deletions, and
modifications through intuitive highlighting, making it easy to spot
differences at a glance.

Deletions are visually highlighted in red, while additions are
distinguished with green color. The right code panel portrays the most
recent version, while the left one corresponds to a specific version
that a user selects for viewing.


![diff0editor](https://github.com/Texera/texera/assets/41690429/ac0d7362-b2fc-48c6-9c17-3617b3dea474)

By integrating this advanced Difference Editor into our Version View, we
have elevated the usability and effectiveness of our version control
system, empowering users to review and track changes more effectively.