Fix upstream link for source UDF operator (#2016)
This PR fixes an issue exposed by #1845, where the upstream link is
shared and enforced on the Python side. However, it did not handle
source operators properly as they do not have a valid upstream link.
This bug caused source (1-out) UDF to fail to start.

This PR proposes a fix by recognizing the case where the upstream link
is None and treating it as a special commendation for source operators.