Fix optional field displaying "no longer a valid option" error message (#2101)
This PR addresses an issue in the current master where certain operators
contain an optional field annotated with `@AutofillAttributeName`. In
the backend code, this optional field is assigned a null value. However,
during schema propagation, the frontend performs a check to ensure that
the
value corresponds to a valid attribute name. The null value will be
converted to an empty string by the frontend, then an error message is
generated for the optional field, as an empty string is not present in
the enum of all attribute names.
<img width="344" alt="screen"
src="https://github.com/Texera/texera/assets/13672781/ad545c43-a150-4606-816a-aa9556ceb8ef">

This PR adds an empty string to the enum list if the field is not marked
as required.

---------

Co-authored-by: Yicong Huang <17627829+Yicong-Huang@users.noreply.github.com>