Add Gantt Chart Visualization Operator (#2099)
A [Gantt chart](https://en.wikipedia.org/wiki/Gantt_chart) is a type of
bar chart that illustrates a project schedule. The chart lists the tasks
to be performed on the vertical axis, and time intervals on the
horizontal axis. The width of the horizontal bars in the graph shows the
duration of each activity.
The operator takes in input data with mandatory fields like task name,
start datetime and finish datetime and produces the gantt chart. It can
optionally take a color field.

![ganttchart](https://github.com/Texera/texera/assets/117004801/4be9a1fc-04f3-429d-ab87-13f6a16546c0)

---------

Co-authored-by: raj-numerade <raj@numerade.com>
Co-authored-by: Shengquan Ni <13672781+shengquan-ni@users.noreply.github.com>