Add TimeSeries Operator (#2112)
This PR introduces a new visualization operator: Time Series

Time series visualization is the graphical representation of data that
is ordered over time. It is essential for analyzing and interpreting
temporal patterns, trends, and anomalies in various fields, including
finance, economics, climate, stock market analysis, etc.

The users need to specify a value column that represents the value that
changes over time, a date column that represents the date value for each
data entry, their corresponding x and y labels, and a tick value for the
plot. The users can also specify the title of their plot if needed.

The users can see the data value and its corresponding date by hovering
their mouse over the plot.

Example of visualization:

![Demo](https://github.com/Texera/texera/assets/61521613/31fd6143-891c-4bf2-a68a-9c461bfa2ead)