Update PieChart Visualization Operator (#2088)
This PR update the a visualization operator: PieChart

The Pie Chart is a classic and widely used visualization that
effectively displays proportions and percentages within a dataset. It is
particularly useful for representing categorical data and showing the
distribution of values in a whole.

The user needs to specify a value column, a name column and a title for
the PieChart. The operator automatically calculates the percentage of
each category and visualizes the proportional slices accordingly.

Example of visualization:
![Jul-24-2023
20-34-59pie4ban](https://github.com/Texera/texera/assets/125719918/d2f4b946-e4fd-438b-bb5e-3870a79a16e8)

---------

Co-authored-by: Yicong Huang <17627829+Yicong-Huang@users.noreply.github.com>
Co-authored-by: Shengquan Ni <13672781+shengquan-ni@users.noreply.github.com>