Add DotPlot Visualization Operator (#2087)
This PR introduces a new visualization operator: DotPlot

Dotplot is a versatile and intuitive way to display data, particularly
when dealing with categorical variables. This visualization method
involves using dots to represent individual data points and their
frequency. The dots are placed along a single axis, with their position
indicating their category and the number of dots at each position
representing the count or frequency of that category.

The user needs to specify a value column for the counting of attributes.

Example of visualization: 

![DotPlot](https://github.com/Texera/texera/assets/110846941/411511ad-ffe4-49d3-b929-c0a097c429f8)

---------

Co-authored-by: Shengquan Ni <13672781+shengquan-ni@users.noreply.github.com>