Add Filled Area Plot  Visualizer Operator (#2086)
This pull request introduces a new plot visualizer, which is Filled Are
Plot.

Filled Area Plot is similar to line plot, but it fills the area under
the line with colors and symbols to indicate the volumn. If there are
multiple lines in a plot, the area under each line can be in different
colors. In such way, the users could visualize and compare different
attributes effectively.

The users need to specify the X attribute and Y attribute to build a
basic plot. LineGroup attribute enables the users to separate the lines.
And other two options are used to make the visualization better:
coloring the plot or splitting the plot.



![demo_gif](https://github.com/Texera/texera/assets/86388854/6a52c93b-62d8-4889-8a65-4d02e29df00a)

---------

Co-authored-by: MiuMiuMiue <yuanzhe2001@gmail.com>
Co-authored-by: Shengquan Ni <13672781+shengquan-ni@users.noreply.github.com>