Add Image Visualizer Operator (#2111)
This pull request proposes the addition of an Image Visualizer to the
Texera workflow system. The Image Visualizer is designed to display
images directly without the need for any conversion. It takes the binary
data of an image as input and renders the image directly in the
visualization output.

Example of visualization:
![Aug-01-2023
21-05-04ImageViz](https://github.com/Texera/texera/assets/125719918/a7545285-c94a-446d-84ab-e51c16598e08)

---------

Co-authored-by: Shengquan Ni <13672781+shengquan-ni@users.noreply.github.com>