Zombie process self-cleanup logic with heartbeat (#2100)
This PR is designed to further address the issue of zombie processes,
particularly focusing on Python processes. We've identified two main
scenarios that lead to the persistence of such processes:

1. Joblib, a popular Python library known for doing parallel processing
and is also used by some popular ML library, will cause zombie processes
left in the system even the code completes the execution. This issue has
been reported for a long time and still remains unresolved.
2. If the parent of Python processes terminates unexpectedly, the Python
processes and their children processes will still exist in the system
and occupy many memories.

In order to solve the issues above, new logic is introduced. First,
after everything is finished, we will do the OS call to clean every
subprocess, which can solve the Joblib zombie process issue. Second, we
have added a heartbeat feature with a 5-second intervals, using async io
to minimize the cost. After the heartbeat fails, all the children
processes as well as the Python worker process will be killed to make
sure the cleanness of OS. Noticed that during this step, we will also
compare the parent PID with the original parent PID, logically it should
be changed when heartbeat failed. Such a result will be written into a
"critical" log that can help us to review the issue, which will be only
created at the time that unexpected issues happen.

A small demo to mock the JVM crashed.
![Peek 2023-08-07
20-41](https://github.com/Texera/texera/assets/114192306/f75a52bd-248f-4167-b63d-c24ab70075a2)

---------

Co-authored-by: Yicong Huang <17627829+Yicong-Huang@users.noreply.github.com>