Execution Dashboard (#2130)
### Introduction
---
This PR introduces Execution Dashboard under the Admin tab in texera,
which allowes the admin to see the process of all executions. If there
are multiple executions running at the same time, the Admin is able
check the state of all executions by checking the execution dashboard.
This dashboard allows the Admin to:

- View list of executions with their information
- Sort the executions according to each column attribute
- Filter the executions according to the status
- Kill / Pause / Resume a running execution
- View history of a specific workflow

### Solution
---
The main goal of this dashboard is to help Admin check as mush
information of each execution as possible and also manipulate the
executions. Each column of the table contains different information of
each execution:

- `Workflow (ID)`:The corresponding workflow name and its ID in
parenthesis.
- `Execution Name (ID)`: The execution name and its ID in parenthesis. 
- `Initiator`: The name of the user who starts the execution.
- `Status`: The status of the execution: Completed, Just Completed,
Killed, Failed, ...
- `Time Used (hh:mm:ss)`: The total time used to run the execution in
the formatof a timestamp.
- `End Time`: The time when the execution finished running in the format
(Date, Timestamp, Time zone).
- `Action`: Actions the user can take to the execution
  - `Kill`: Kill the running execution immediately.
  - `Pause`: Pause the execution immediately.
  - `Resume`: Resume paused execution.
  - `History`: View the history of the corresponding workflow.

In addition to the Execution Dashboard web interface, this PR also
modified the `aborted` status to `failed` and added the status `killed`
for killed execution.

Current PR contains the backend codes, and the frontend codes are in
this PR #2131

### Demo
---


https://github.com/Texera/texera/assets/86388854/a44550ba-a8db-40f8-b1ae-7af18f8f285a

---------

Co-authored-by: MiuMiuMiue <yuanzhe2001@gmail.com>