Add back operator caching feature  (#2154)
This PR adds a new implementation of the operator caching feature under
the new compiler framework, there are two major changes:

1. In the backend, added a new `WorkflowCacheChecker` implementation,
which uses equivalence classes to detect changes in a workflow to decide
if the last execution can be reused.
The new implementation is only ~100 lines of code, which is
significantly simpler than the original implementation (~500 lines of
code).

2. In the frontend, separarted the concept of `view result` of an
operator, and `reuse the cached result` of an operator, these are now
two buttons that the user can control separately. The user can even add
`reuse the cached result` after the last execution and before the next
execution, which is more user-friendly.



### View result of intemediate operator
Click "view result" or use the right click context menu. Results of
intemediate operators are also shown to user.

![viewresult](https://github.com/Texera/texera/assets/12578068/898db8ac-991f-42f9-b279-2f688a30e87b)



### Reuse cached result of an intemediate operator
Click "reuse result" or use the right click context menu. Only the
operators after the "reuse cache" are run.


![reuse_result](https://github.com/Texera/texera/assets/12578068/1ed6b948-c499-4475-a76c-e7224d41f571)


### Modifying workflow invalidates previous cache
If the content of the cached operator is modified, the previous cached
results will be invalidated, and the icon is updated to indicate the
cache cannot be used.


![invalidate_cache](https://github.com/Texera/texera/assets/12578068/e2a9b40e-ec11-4ee6-ae8d-cc49c02288be)

Modifying the content of upstream operaotrs also invalidates the cache


![invalidate2](https://github.com/Texera/texera/assets/12578068/8d780913-7686-48ed-8cec-242f1c22510c)