Fix flow control first message issue (#2202)
This PR fixes an issue with the current flow control mechanism. When a
worker recovers from a backpressure pause, it mistakenly sends out a
stashed message regardless of its size. Now it will check the available
credit and only send out a stashed message that is within the available
credit.