Fix port descriptor (#2229)
This PR fixes an issue where a port described on the frontend does not
reflect on the backend. This is because we previously only allowed
dynamic ports to be describable.

In this PR, we generalize the setting and let every port description
from the frontend be sent to the backend. And let all backend operators
take and parse the port descriptions (by extending `PortDescriptor`).