Use ActorMessage for flow control (#2237)
### Introducing a new type of message: `ActorMessage`
This message is in parallel to `DataMessage` and `ControlMessage`.
ActorMessage each contains an `ActionCommand`, whose handler has the
following properties:
1. It will be handled by the receiver (e.g., `NetworkReceiver` on the
Python side).
2. It does not access or alter any data processor's state. It can only
access, including read and write, the input queue. The operations are:
    - enable/disable the input queue;
    - read from the input queue (e.g., get credit);
    - create and put a headless Control Message to the input queue.

### Flow control with ActorMessage:
1. Backpressure implementation:
`Backpressure` is now an `ActorCommand`. The handler will disable the
data subqueue of the input queue to realize the operator being
pressured. This implementation also pressures the source operators
(e.g., PythonSourceOp) correctly. We introduced a `NoOp` Control command
to resume the Python main loop when backpressure is lifted.
2. PollCredit implementation:
`CreditUpdate` is now an `ActorCommand`. The handler does nothing but
return the latest queue credit back in its ack, in a synchronized
manner.

### Other applications of ActorMessage:
ActorMessage is designed for messages that are to be handled by the
actor without modifying or altering the Data Processor's state. Besides
flow control, it can also be used to implement TimeService messages in
which the actor needs to send itself a message to trigger some actions.
This is to not be confused with `ControlMessage` since control messages
will be logged to ensure fault tolerance.

### ActorMessage's fault tolerance and replay.
ActorMessage should not be logged and does not need to be replayed.

---------

Co-authored-by: Shengquan Ni <13672781+shengquan-ni@users.noreply.github.com>