Refactor Python UDF initialization (#2236)
This PR aims to provide an operator interface to support operators that
are designed to be compiled later. This is particularly useful for UDF
operators. In this PR we refactor the existing Python UDF operators onto
this new interface. Future PRs will add other languages to use this
interface to define an operator executor.

API to initialize with an IOperatorExecutor: (existing java/scala native
operators are implemented with this API)
```
OpExecInitInfoWithFunc(opGen: ((Int, OpExecConfig)) => IOperatorExecutor)
```

API to initialize with a code String: (Python UDF and native operators
are implemented with this API)
```
OpExecInitInfoWithCode(codeGen: ((Int, OpExecConfig)) => String)
```

---------

Co-authored-by: Shengquan Ni <13672781+shengquan-ni@users.noreply.github.com>