Refactor replay for fault tolerance (#2231)
This PR reimplemented the replay part of Amber for Fault tolerance.

## Overview of Replay:
To support the replay of the processing of an actor. We need to enforce
the processing order of input to be the same as the logged order. The
replay is disabled now, will be enabled after later PRs.

### ReplayGatewayWrapper:
If the worker is in replay mode, we will wrap the `InputGateway` with a
`ReplayGatewayWrapper`. `ReplayGatewayWrapper` uses the normal
`InputGateway` to enforce FIFO, but it enforces the logged order of
message processing. After the replay finishes, the wrapper will
disappear, switching the actor to its normal behavior.

### The replay process:

![Texera Overall Infrastructure-refactored worker drawio
(2)](https://github.com/Texera/texera/assets/13672781/212b0317-9148-47c8-b503-8270cb65aa1f)

---------

Co-authored-by: Yicong Huang <17627829+Yicong-Huang@users.noreply.github.com>