Rename logical level entities (#2257)
## This PR refactors the LogicalPlan with code clean up and renamings,
without functionality changes.

### Renames the logical-plan-level entities (old -> new):
1. `OperatorDescritor` -> `LogicalOp`
2. `OperatorLink` -> `LogicalLink`
3. `OperatorPort` -> `LogicalPort`

### Redesignes Identities (old -> new):
1. `OperatorIdentity(workflow: String, operator: String)` ->
`OperatorIdentity(id: String)`: We have been using the `workflow` in
`OperatorIdentity` as the executionId. It is redundant as we also pass
executionId from `WorkflowIdentity` (see later). This change removes the
redundant field.
2. `WorkflowIdentity(id: String)` -> `WorkflowIdentity(executionId:
Long)`: We have been using `id` in `WorkflowIdentity` as the
executionId, this rename reflects the actual usage of the field.
3. `LayerIdentity(workflow: String, operator: String, layerID: String)`
-> `LayerIdentity(operator: String, layerID: String)`: This change
belongs to the physical layer, we removed `workflow: String` due to the
removal of that from `OperatorIdentity`.

### As a result, the Logical Plan is changed to the following:
Old:
```
operatorMap: Map[String, OperatorDescriptor] // where String is the operator Id;
jgraphtDag: DirectedAcyclicGraph[String, OperatorLink] // where the String is the operator Id.
```
New:
``` 
operatorMap: Map[OperatorIdentity, LogicalOp]
jgraphtDag: DirectedAcyclicGraph[OperatorIdentity, LogicalLink]
```