Add URI and folder support to the storage of replay log (#2254)
To manage the lifecycle of the replay log better, this PR redesigns
the`logStorage`.

### The Problem 
ach log file is independent and identifiable solely by its name.
Highlight that currently, each `logStorage` instance is linked to just
one log file. In a single execution, we may create many workers, which
generate multiple log files, leading to a fragmented and worker-centric
log management system.

### The Solution
I made the `logStorage` manage a folder for each execution of a workflow
instead of individual files for each worker. This approach centralizes
log management at the execution level.

### Use URI to identify the folder location
The folder for each execution is represented as a URI. This ensures
file-system independence and facilitates easy switching between
different file systems like Windows, Mac, or HDFS for log storage.

---------

Co-authored-by: Yicong Huang <17627829+Yicong-Huang@users.noreply.github.com>