Refactor backend ports in compiler (#2299)
This PR refactors the ports on the backend.

### PortIdentity
Previously, we store ports in a list and used the index to refer to a
port. This is not useful when the ports are mapped from logical operator
to physical operator (the exact sequence of the ports may change), or
when supporting dynamic ports (removing a port would cause the index
change).

Now we introduce `PortIdentity` as the id of a port. The same port and
identity will be used across logical, physical, and in workers.

The exact structure of PortIdentity is shown below (generated by
protobuf):
``` 
final case class PortIdentity(
    id: Int = 0,
    internal: Boolean = false
)
```
A PortIdentity has a field for indicating `internal` or not. This is to
differentiate internal ports and external ports. Internal ports are
ports that are added between the physical operators of a logical
operator. External ports are those set on the logical operator and link
to other ports from other logical operators.

### New InputPort and OutputPort
The `InputPort` has been updated (generated by protobuf):
```
final case class InputPort(
    id: PortIdentity = PortIdentity(),
    displayName: String = "",
    allowMultiLinks: Boolean = false,
    dependencies: Seq[PortIdentity] = Seq.empty
 )
```
Notably, we moved `dependencies` from the PhysicalOp into InputPort. It
is now a sequence of PortIdentities which denotes the ports that this
port is depending on.

The `OutputPort` has minor changes and we will adding the `blocking`
property into it in the future.

### LogicalLink and PhysicalLink
Both links are updated to use PortIdentities instead of port index (Int)
accordingly:
```
case class LogicalLink(
    @JsonProperty("fromOpId") fromOpId: OperatorIdentity,
    fromPortId: PortIdentity,
    @JsonProperty("toOpId") toOpId: OperatorIdentity,
    toPortId: PortIdentity
)
```
```
final case class PhysicalLink(
    fromOpId: PhysicalOpIdentity,
    fromPortId: PortIdentity,
    toOpId: PhysicalOpIdentity,
    toPortId: PortIdentity,
)
```

### Frontend ports plan
For now, we add some adapt functions to help hookup with the new backend
design, and fully supports backward compatibility.
 
Future plan:
- The frontend has other redundant definitions such as `PortDescription`
and `PortProperty` which can be further merged and refactored.
- Also, we need to hookup `dynamic port` with the new backend. 

### In-worker ports plan
Workers (both Python and Java) are using ports as well. Right now they
are still using port index, or `PortOrdinal`. We plan to unify them in
the future as well.