Support multiple sources in a Region (#2314)
As we are doing region-by-region propagation to generate resources, a
region needs all its links to be ready for partition requirement
propagation. For an operator that has multiple sources connected to
multiple ports, if there is no dependency between the input ports, then
all sources can be within the same region. This PR changes the region to
have multiple sources and removes the old single-source limitation.

We don't currently see issues stated in #1768. We will fix them if
encountered.