Refactor WorkflowScheduler to be WorkflowExecutionController (#2372)
This PR introduces `WorkflowExecutionController` as a drop-in
replacement for the current `WorkflowScheduler`. This is aligned with
our direction to have a new `WorkflowScheduler` later to decide region
plans.

The change mainly includes the following two new components:
1. WorkflowExecutionController: It is a component inside the Controller,
which manages the execution of a given workflow.
2. RegionExecutionController: It manages the execution of a given
Region. The execution status (`RegionExecution`) is currently stored
inside each RegionExecutionController, this will be refactored with
other execution status (`Execution`) classes later.

We also removed `SchedulingPolicy`, as it should be an input to the
RegionPlanGenerator. Currently, we use the default case where multiple
Regions with the same rank can execute at the same time.