Migrate Tuple definition from Java to Scala (#2417)
This PR transitions the Tuple definition from Java to Scala. It's
important to highlight that the usage of Object has been adapted to Any
in the Scala implementation. This change is pivotal because Scala's Any
encompasses a broader spectrum than Java's Object, primarily due to
Scala's native support for unboxing primitive types. In the context of
Scala, Object is analogous to AnyRef, which specifically refers to
reference types. This adjustment ensures a more seamless integration of
primitive and reference types within Scala's type system, leveraging
Scala's capabilities for more efficient and type-safe code.