Add Dummy Operator and Dummy Property (#2363)
This pull request introduces a new property, the dummy property, and a
new operator, the dummy operator. These new features facilitate the
process of workflow translation from a 3rd party application to Texera
by acting as placeholders for unimplemented operators and features in
Texera.

The dummy operator is a placeholder for operators that exist in a 3rd
party application but do not exist in Texera. Its main purpose is to
maintain the DAG structure of a workflow after it undergoes translation.

The dummy property is a placeholder for operator fields that don't exist
in Texera. It can be used in conjunction with the dummy operator to
describe its intended fields and properties, as well as with
pre-existing operators in Texera that have an equivalent 3rd party
operator. Its purpose is to ensure that information about an operator
isn't lost after it undergoes translation.

These features are intended to be read only; users won't have the
ability to place or edit dummy operators or dummy properties. Instead,
these features are to be used with a workflow parser, which translates
3rd party workflows to Texera workflows stored in JSON format.

Below are two brief demos showing how the dummy property and dummy
operator appear in Texera.


https://github.com/Texera/texera/assets/70456187/35b31d63-3065-4d8b-9ec6-3018029be157



https://github.com/Texera/texera/assets/70456187/01beef39-a817-4aee-afd3-82ba73529f9e

---------

Co-authored-by: Xiaozhen Liu <xiaozl3@uci.edu>