Add Environment GUI (#2444)
This PR introduces the GUI of `environment` and some fixes to previous
`dataset` features. For the backend of `environment`, see #2434

### Features

- Environment Tab on the left panel
![2024-03-04 23 26
19](https://github.com/Texera/texera/assets/43344272/09738edd-aa91-4a1f-a915-4d41f25afc9b)

- Auto Complete only from the files in datasets
![2024-03-04 23 27
32](https://github.com/Texera/texera/assets/43344272/36178dea-24fd-41f3-92e2-800c10f21ae5)

### Implementation Details

1. The changes on the `ScanSourceOperatorDesc`

Previously, the source file is located by its absolute path and scanned
into the workflow. Now, since all the files are within the dataset and
managed by JGit, its physical file may not be directly available,
current solution is to write the target file into a temporary file,
which is identified by an absolute path generated by JVM. The file will
be deleted when JVM quits.

2. When workflow execute request is submitted, the webserver will also
persist the environment eid to the `WorkflowExecutions` table.