Remove the Dependency of Environment when Workflow executes without user system enabled (#2452)
This PR fixes the issue that when user system is enabled, the workflow
execution still depends on the environment and MySQL.

## How the fix is done

Since when user system is enabled, the workflow execution will be stored
in the MySQL, along with the `environment_eid` which indicates which
environment the workflow execution happens. I moved the fetch of
`environment_eid` to the
`ExecutionsMetadataPersistService.insertNewExecution`, only when user
system is enabled will the environment_eid be fetched.

## Some consideration

Regarding #2451 , since the whole dataset+environment things rely on
MySQL, and `user-system-enabled` flag indicates whether to use MySQL, in
this fix I didn't include another flag to enable/disable environment, as
this flag has dependency with `user-system-enabled`, meaning that this
`environment-on` flag will only work if `user-system-enabled` is on. I
think we can discuss and introduce this flag in the future development.