Initialize executors through message (#2446)
This PR changes the life cycle to initialize an executor on Java/Scala
side. Previously, we sent the PhysicalOp through the Worker (actor)
initializing phase, using `preStart` hook to do the executor
initialization. This exposes an issue where if there is any error during
executor initialization, the worker node JVM would crash. Handling such
an error will surpass the exact cause and let a worker create without a
valid executor, and it would later throw other exceptions such as
null-pointer exceptions.

Now, we let the Worker (actor) create without initializing the executor.
We send in the OpExecInitInfo through a control message
`InitializeOperatorLogic`. This is aligned with the Python operator's
initialization life cycle, in fact, now they are using the same control
message.

As a result, this PR also removes the requirement on PhysicalOp during
the initialization of a Worker. Such a requirement is not necessary.

As a next step, we will align the `UpdateOperatorLogic` with the Python
side to update an executor instance during the execution.