Add TablesPlot Visualization Operator (#2461)
This pull request introduces the Tables Visualization operator.

Table object for detailed data viewing. The data are arranged in a grid
of rows and columns. Most styling for headers, columns, rows, or
individual cells can be specified. The table uses a column-major order,
ie. the grid is represented as a vector of column vectors.

For implementation details, this operator is implemented as a
Python-native operator. It accepts user-provided data and generates the
corresponding Python script to create Tables. The visualization is
created using Plotly.

For further information on utilizing the Tables functionality within
Plotly, please refer to the
[documentation](https://plotly.com/python/table/).


![TablesDemo](https://github.com/Texera/texera/assets/42431493/53ca0a7e-e146-472b-b832-e61ee91e55cb)

---------

Co-authored-by: Harsh Sanjeev Agarwal <hsagarwa@uci.edu>
Co-authored-by: Xiaozhen Liu <xiaozl3@uci.edu>