Rename Operator to Executor in workers (#2460)
The name "operator" has multiple meanings in the system. To avoid
confusion, we now name the component that runs in a worker as
`Executor`. This also changes all the corresponding managers, control
messages and function names.

For backward compatibility issue, for Python side, we keep the Operator
class for now.