Support partial workflow execution up to an operator (#2259)
Introducing a new feature called "execute up to this operator" (related
to #1950). Here is a small demo where user can run a partial DAG of the
workflow by selecting which operator to run to. This can be beneficial
when user has some part of the workflow that’s not valid yet (e.g., has
not finish constructing).

Currently, we provide this feature through context menu option. In the
future, we can have other designs, for example, a play button can be
shown on each operator for execution till this operator.

It co-exists with the global workflow run button. Gradually we may
remove the global run button.

### Demo
User can select which operator to run to.
![2023-12-10 21 20
15](https://github.com/Texera/texera/assets/17627829/480d84a0-7554-4bf7-a2ad-58abdeecbb5d)

With cache enabled:
![2023-12-10 21 30
22](https://github.com/Texera/texera/assets/17627829/f4d55a13-1ac1-4cd5-9350-639023fc8e36)