Clear the workflow metadata when workspace view destroy (#2498)
This PR resolves the issue that:
when switching from workspace of workflow with wid `a`, to workspace of
workflow with wid `b`. wid `a` is showing up when components in the
workspace(e.g. left-panel) try to retrieve the wid from the
workflowActionService.

## How the fix is done

Add one line in the `ngOnDestroy` of `workspace.component`, to set the
workflowMetadata as the `DEFAULT_WORKFLOW`