Fix The Issue That Python Operators Cannot Be Paused And Resumed (#2590)
This pull request addresses an issue preventing Python operators from
properly pausing and resuming. The problem was traced back to a
modification made in PR #2570, where the return type of the
ControlReturnV2 message no longer included WorkerState. As a result, the
async RPC server was unable to process the message. This PR fixes the
issue by adding WorkerState back as a valid return type for
ControlReturnV2. Now, both the pause and resume handlers will return a
message that reflects the worker's current state.