Grant Dummy Operators Access to Dummy Property Field (#2572)
This is a minor change that allows dummy operators, from the workflow
parser, the ability to access the dummy properties field. In the future,
the workflow parser will take advantage of the dummy property feature to
describe properties of equivalent operators from other workflows that
aren't currently present in Texera. Below is an image of what this
functionality looks like.


![Capture](https://github.com/Texera/texera/assets/70456187/0c8a0b13-4e2c-48d2-b4d7-f3de52af4f0e)

Co-authored-by: Xiaozhen Liu <xiaozl3@uci.edu>