Fix Result Panel Not Showing Results Occasionally (#2623)
This PR fixes an issue of result panel that could cause results not
being shown from time to time (#2612 ). The issue is caused by #2604.
The `currentPageIndex` should start from 1 but #2604 introduced a logic
that could change the index to 0. Adding a check in the condition could
solve this issue. I tested and did not see problems anymore.