Fix blinking issue in result panel during the execution (#2690)
Since the backend sends "delta" updates of the results, there is a
chance that nothing has changed between the two result updates. Thus,
the backend may send an empty update to the frontend. This PR removes
the logic that clears all results when an empty WebResultUpdate event is
received, ensuring the user can see the results during execution.