Enhancing Report Generation by adding Operator Results (#2792)
This PR enhances the report generation functionality by integrating
comprehensive operator results handling, allowing for more detailed and
informative reports. This PR is a continuation and enhancement of the
previous PR, which can be found at the following link:
https://github.com/Texera/texera/pull/2770

The file was modified in this PR: ReportGenerationService
getAllOperatorResults(operatorIds: string[]): Collects all operator
results from the workflow, processes them, and generates an HTML report
containing both the workflow snapshot and the results for each operator.
The final report is automatically downloaded with a file name derived
from the workflow name.

retrieveOperatorInfoReport(operatorId: string,allResults: { operatorId:
string; html: string }[]): Fetches and processes the result for a given
operator, generating an HTML snippet that is included in the final
report. Handles different result types, including paginated results and
snapshot results.

generateReportAsHtml(workflowSnapshot: string, allResults: string[],
workflowName: string): Generates a comprehensive HTML file containing
the workflow snapshot and all operator results, then triggers a download
of the report. The file name is created using the workflow name,
ensuring easy identification of the report.

The file was changed in this PR: MenuComponent
onClickGenerateReport: Displays a notification indicating that the
report generation has started, collects the operator results using
getAllOperatorResults from ReportGenerationService, and then calls
generateReportAsHtml to generate and download the comprehensive HTML
report.

Operation Process: Click the button below to generate the report with
detailed operator results and the workflow snapshot.

![image](https://github.com/user-attachments/assets/38311e39-cda1-4c85-a4f2-80d4f4a092fc)

Here is a part of the example report.

![image](https://github.com/user-attachments/assets/e93b10b4-347a-4ce1-aa8b-45d943b56c21)

![image](https://github.com/user-attachments/assets/128aa56a-867f-4ee3-a15e-8f7fe32a5cd5)