Freeze Dependency For Texera (#2897)
**Description**: This Pull Request aims to freeze all direct
dependencies for the Texera project to reduce potential bugs caused by
automatic dependency updates. By locking the versions of all
dependencies to the currently tested and stable versions, we can improve
the reliability and consistency of project builds.

**Rationale**: Frequent updates to dependencies may cause unforeseen
compatibility issues, leading to project instability or unexpected
errors. By freezing dependency versions, we aim to prevent such issues
and allow for gradual, controlled updates of dependencies when necessary
in the future.