Enable Bulk Download of Binary Data from Result Table (#2931)
This PR introduces a feature that allows users to perform bulk downloads
of binary data from the result table of a general operator. The
motivation stems from feedback during a demo of the CloudMapper
operator, which highlighted the need for more convenient data export
functionality. The CloudMapper operator often produces multiple rows,
with each row containing three binary data entries.

Previously, users could only download individual binary data cells one
by one, which became cumbersome when dealing with large datasets. With
this new feature, users can download all binary data files in a single
click, enhancing the usability and efficiency of the interface.

### Overview:
- Bulk Download Button:
  - Located in the result panel.
  - Appears only when the result table contains binary data.

- Download Process:
- On clicking the button, all binary data from the result table is
downloaded as a ZIP file.
  - Structure of the ZIP file:
    - Folders represent rows in the result table.
- Files inside each folder correspond to binary data entries in the
columns of that row.
    - Each file is named after the column it represents.
    
### Known Issue:
There is a bug in the frontend function responsible for checking if the
data schema contains binary data. Although the backend has access to the
correct schema, it is not being propagated properly to the frontend.
This issue will be addressed in the next PR. In the meantime, the
existing frontend check is being used to determine if the data is
binary.


https://github.com/user-attachments/assets/d60e5078-20b7-48f3-975c-76534a9181bb

---------

Co-authored-by: Kunwoo Park <kunwoopark@Kunwoos-MacBook-Pro.local>
Co-authored-by: Kunwoo Park <kunwoopark@dhcp-172-31-184-236.mobile.uci.edu>