Add nz-spin Loader to Workspace (#2966)
This PR introduces an `nz-spin` loader to the workspace to indicate that
a workflow is being loaded. Previously, when uploading a large workflow,
the workspace would appear blank during the loading process, leaving
users unsure of what was happening. With this change, the `nz-spin`
loader ensures users are aware that the workflow loading is in progress,
thereby creating a more user-friendly experience.
<img width="1433" alt="Screenshot 2024-10-24 at 3 55 49 PM"
src="https://github.com/user-attachments/assets/d76a388c-a15d-48c6-9066-73dd2619cfd7">


https://github.com/user-attachments/assets/3888a5f8-a12c-40f3-9140-ed383db7a3c6

Co-authored-by: Kunwoo Park <kunwoopark@dhcp-172-31-218-249.mobile.uci.edu>