Fix avatar display issue caused by excessive requests (#3118)
### Purpose:
Currently, all avatars are fetched via their URLs. When many identical
avatars are displayed, such as in the user workflow interface, excessive
requests may cause them to fail to load. This PR resolves the issue by
caching avatars with the same URL to prevent redundant requests.

### Changes:
1. Added caching to store previously loaded avatars with a retention
time of 1 hour. All caches are cleared when the user logs out.
2. In the User Avatar component, the cache is checked for available
avatars. If an avatar is found, it is used directly.

### Demos:
Before:

![image](https://github.com/user-attachments/assets/724fb70b-703c-49c4-a0a7-80db3797a8a7)

After:

![image](https://github.com/user-attachments/assets/6eb38d36-9b3e-40ec-b0e3-388e05ba65ec)

---------

Co-authored-by: Xinyuan Lin <xinyual3@uci.edu>