Refactor APIs related to Hub interaction activities (#3252)
### Purpose:
The current different options, including workflow and dataset, have APIs
related to Hub interaction activities such as like, clone, and viewCount
that cannot be shared. This results in a large amount of redundant code.

### Changes:
1. **Update the `activity` table** by dropping the existing
`workflow_user_activity` table and create a new `user_activity` table,
adding the `type` attribute to support storing different types of
activities. Please use `20.sql` for the update.
3. **Rename `HubWorkflowResource` to `HubResource`** to clarify that the
logic within can be applied to all types of Hub activities.
4. **Move APIs directly related to workflows** into `WorkflowResource`,
retaining only activities that are directly related to the Hub.
5. **Refactor the APIs in `HubResource`** to accept an additional
attribute `entityType`, ensuring that multiple options (e.g., dataset
and workflow) can share a common API.
6. **Add `EntityTables.scala`** to manage different entity tables in a
unified manner. In the future, if a new option such as `Project` needs
to be added, it can be done by simply adding a corresponding `case
object` in this file without modifying `HubResource`.