Fix the issue where buttons around the operator are difficult to click. (#3277)
### Purpose:
Currently, buttons around an operator (e.g., add port, remove port) are
very hard to click. This is because the buttons are shown when the
cursor is "over" the operator, and will disappear when user try to move
the cursor towards the button.
Fix #2269

### Changes:
1. Remove the effect where buttons only appear when hovering over the
operator and disappear when moving to a blank area.
2. Add functionality so that buttons appear only when the user clicks or
right-clicks on the operator, and disappear when the user clicks on a
blank area or another operator.

### Demos:


https://github.com/user-attachments/assets/b8ef6c7e-5f7d-4636-9c72-823324a6ced5