Fix the issue where the workflow cannot be displayed in the user project. (#3283)
### Purpose:
Currently, after adding a workflow to a project, the workflow list
cannot be viewed within the project. The issue is caused by the workflow
display interface within the project not occupying the full height of
the browser window, resulting in the section displaying the workflow
being compressed and thus unable to show the workflows in the project.

### Changes:
Modify the styles in the project to make it occupy 100% of the page
height and ensure that the workflow results take up the remaining space.

### Demo:
Before:

![image](https://github.com/user-attachments/assets/9a2ce40f-a3f4-4589-82fc-9d1f877c7292)

After:

![image](https://github.com/user-attachments/assets/9d672440-5d7f-48a2-8723-439157e02b41)