Add Layer Info in Port Result Identifier (#3279)
## Motivations

Currently the operator-port-result storage identifier is based on the
logical opeator id only. This is problematic, especially when it comes
to scheduler-generated materialized ports. For example, if the scheduler
creates a materialization on the the output port of the `localAgg`
physicalOp in an `Aggregate` operator, and this logical Op happens to
need view results on the frontend, the two storage identifiers will
conflict.

To solve this issue, and also to allow a physical operator port to use
the same storage for both view result case and the scheduler-added
materialized result case in the near future, this PR updates the current
operator-port-result storage identifier to include `layerName`, which
combined with the `operatorId` (`logicalOpId`) can be used to identify a
specific physical operator.


## Designs
- Both the URI and the psql schema are modified to include `layerName`
as part of the identifier.
- When creating a storage in the compiler, the `layerName` of a
physicalOp is added as part of the identifier for the document.
- When accessing the storages of a logical operator for processing
frontend requests, as the `layerName` is not available, we retrieve all
the available URIs associated with a logical operator, and return only
the first URI that satisfies both these two conditions:
1. the URI is for an external output port (e.g., the output port of
`globalAgg` in `Aggregate` or `probe` in `HashJoin`)
2. the URI is a `RESULT` type as opposed to a `MATERIALIZED_RESULT`
type. (Note this will be unified soon)

- The scheduler will generate materialized result objects using the
specific `layerName` of a physical operator.

## Related bug Fixes
This PR also fixes two related bugs introduced by recent codebase
changes.
1. #3264 fixes the issue of duplicate sink generation for the same
materialized result read by two downstream operators. However the
multiple downstream operators still need different cacheRead operators
to avoid being scheduled in the same region. Currently the two cacheRead
ops will have the same id, which ultimately makes them the same
operator. This PR adds this fix by including the downstreamOp's id in
the cacheRead's id.
2. There have been many issues related to sink operator's completion
logic. Currently a sink is determined to be finished when its input port
finishes, but that will cause indeterministic problems both on the
frontend status update and on the storage (for example, a materialized
result might be read by the next operator before it is ready because the
writing operator has not finished committing the tuples but is reported
as finished.) This PR adds a temporary fix by moving the result commit
logic to when the input port of the sink completes, which solves the
storage issue. Note the sink operator will be removed in the future,
which should eliminate these problems.


## Applying Database updates to texera-db

Run `core/scripts/sql/updates/01.sql` in psql to apply the changes to
`operator_port_executions` table.