Update Pgroonga stemming filter detection in SQL scripts (#3311)
The previous method for detecting whether the stemming token filter is
present in pgroonga was not functioning correctly in the hub
environment. As a result, I improved the scripts to address this issue.

This PR does not affect the local dev environment as stemming is only
used in the production environment.