Remove Sink Operator in the Backend (#3312)
### Contents of this PR
1. Completely removes sink operators in the backend, including physical
sink operator, sink operator executor and related proto definition, and
places that refer to the sink operator.
2. Replaces the logic of `getPorts` of a region by saving ports
separately (previously we are using links to get the ports).
3. Fixes the timing of flushing storage buffer for port storage writer
threads.
4. Refactors Python worker to send `PortCompleted` even if no links are
connected to an output port.

Relevant test cases are also modified to remove sink operators from the
test cases.

### Regarding region completion and `getPorts`:

Currently our implementation of region completion logic is not correct
as it uses links of a region to get ports of the region and use the
completion of these ports to indicate completion of the region. This
hacky implementation worked in the past when we created sink operators
explicitly but will not work in this PR.

Ideally:
- A region should be defined using only the operators and links in the
region.
- The completion of a region should be based on the completion of its
operators, and the completion of an operator is defined by the
completion of all of its ports.

However, we currently have a hacky implementation of “hash-join”-like
operators that have dependee input ports, and we include operator with a
dependee input port in two regions, which prevents us from relying on
operators to define the completion of a region. Our ultimate goal is to
have a clean separation of regions, but that requires removing cache
read operator first.
To proceed with removing sink operator, in this PR, we implement a
workaround in the region completion logic and do not rely on operators
of a region to get the ports.
After clean separation of regions is done, we can implement the clean
definition of region completion logic and remove redundant port
information in regions.

### Progressive Computation

**Note: This PR removed ProgressiveSinkOp. We will remove
ProgressiveUtils soon due to the high cost of implementing the
retraction operation. Since we now use Iceberg for storage with an
append-only default, removing a record requires a full scan, making
retraction impractical. We will revisit it if there is more use cases
requiring progressive computation.**

### TODOS that are not part of this PR
- Rename "sinkStorageTTLInSecs" and
"sinkStorageCleanUpCheckIntervalInSecs" in the AmberConfig.
- Remove/rename mentions of sink in the frontend.