Fix unaligned representation of python binary objects (#3347)
In #3331, we changed the binary field representation from a single byte
string to a list of byte chunks to support large objects. However, this
introduced an inconsistency in how the data is interpreted across
operators. On the sender side, the tuple field contains the original
Python object, but after transmission, the receiving operator sees the
field as a list of byte chunks instead of the original object. This PR
fixes the bug by ensuring consistent serialization and deserialization
of binary fields across operators.

This PR also fixes a single object being parsed as multiple objects on
the receiver side by appending all the chunks together before parsing.