Fix export in cluster mode (#3378)
## Issue
In cluster setting, computing units are distributed and export request
should be sent to the corresponding computing unit. Currently, we use
generic `execution` endpoint which works on single node (or local)
deployment but does not work with cluster setting.

## Fix
We need to incorporate `cuid` attribute in export request so `envoy`
service forward the request to the correct computing unit.
We replicated `computing-unit-selection.component.ts` logic to receive
`selectedComputingUnit` in `result-exportation.component.ts` and pass it
to `download.service` and finally put it as a parameter in the URL.
![Screenshot from 2025-04-14
14-21-53](https://github.com/user-attachments/assets/bff429ab-8958-4f88-bbbe-6f73f36b76c4)

---------

Co-authored-by: Jiadong Bai <bobbaicloudwithpants@gmail.com>