Revert Changes to Binary Chunking Due to Side Effects (#3431)
This PR reverts PRs #3331 and #3347, which introduced changes for
splitting large binary data into smaller chunks and storing them as a
list within a tuple. The revert is necessary due to several unintended
side effects:

1. Python UDFs receive a list of binary values, requiring users to
implement custom logic to reconstruct the original binary data.
2. As noted by @aglinxinyuan, some control messages are in binary format
and are not handled correctly under the new implementation.