Remove Marker Class (#3467)
This PR introduces no functional changes; it solely involves renaming
classes and removing the deprecated `Marker` class.

`State` and `InternalMarker` remain part of the system, but after this
change, they no longer inherit from the `Marker` superclass, as their
roles differ significantly.

Currently, we still see the `StateFrame`, which contains the `State`
variable as a `DataPayload`. We may change this later.

As for `InternalMarker`, they do not need to be included in any payload.
They are not transmitted over the network and are only visible within
each executor. Their sole purpose is to support the internal logic of
individual operators.