Fix issue where unsaved comments are lost and cursor jumps in the admin user interface (#3464)
### Purpose:
In the admin user interface, when a user tries to add a comment and uses
the **mouse** to reposition the cursor, the cursor always jumps to the
end of the comment text, and any changes the user has made are removed.
However, this issue does not occur when the cursor is moved using the
**keyboard**. The root cause is that the `startEdit` event was bound to
the entire table cell (or an outer container), so when the user clicks
inside the input field to reposition the cursor, the click event bubbles
up to the parent and retriggers `startEdit`. This resets the component’s
state and rebuilds the DOM, causing the cursor to jump to the end and
discarding the unsaved input.
fix #3461

### Changes:
1. Bind the `startEdit` method to the read-only container `<div>` only
2. Add `(click)="$event.stopPropagation()"` to prevent click event
bubbling inside the input field

### Demo:


https://github.com/user-attachments/assets/e9c55b49-6206-438a-b487-181e0f4cc3b9

Co-authored-by: Xinyuan Lin <xinyual3@uci.edu>