Temporary Fix for Python Partitioner Index Issues (#3470)
### Problem

Currently some Python partitioners are not working properly:

- Partitioning information is sent from Java controller to Python
workers as a (ordered) List of Channels .
- When extracting a list of receivers from Channels in Python, currently
we are using an intermediate `{}` (set), which removes the ordered
nature of the list.
- As Hash and Range partitioners are currently using the index of the
list of receivers for their partitioning logic, and the index do not
correspond correctly to the downstream workers, these two partitioners
do not work correctly currently.

### Fix

The fix is to simply avoid using set in these two partitioners.

### TODO

Ideally, we should use a Map for all these implementations as List is
fragile and prune to such issues. We need another PR to update the
`Partitioning` information to be a map.