Add `Parquet` option in export workflow result types (#3478)
## New feature
Users may need to export result of one or multiple operators as Parquet
format. This includes exporting to local or dataset.

## How implemented
In `IcebergDocument` we implemented `asInputStream` as the underlying
function to zip all Parquet files stored in Iceberg.
By implementing this function in `IcebergDocument`, we are now able to
directly access the files stored by storage service. In our case, since
we need Parquet format, we can directly stream back the files saved by
Iceberg without need to convert and consume extra computation.

## Feature behavior
As shown in the picture, now users can select Parquet as export type. 

![Screenshot from 2025-06-18
10-50-54](https://github.com/user-attachments/assets/4db1f6b2-42f5-4ae9-86c9-3170ea03e793)