feat(amber): enable terminating workers in a region upon completion of region execution (#3468)
This PR adds the ability for `RegionExecutionCoordinator`s to terminate
its workers when it detects all the workers are completed.

## Design

- A new control message `EndWorker` is added. 
- When a `RegionExecutionCoordinator` syncs its completion status with
`RegionExecution` and detects it is completed, it sends an `EndWorker`
control message to all the workers of this region and waits for their
confirmations.
- After receiving this final response from all the workers,
`RegionExecutionCoordinator ` sends `gracefulStop` to all its workers,
which terminates all the worker actors (PVMs are also killed
automatically).

## Testing

Use the same workflow and file in #3425 to test this PR. Expectations:

- The workflow should execute successfully without any issues.
- If you set a breakpoint in `WorkflowExecutionCoordinator`, line 86,
and run the workflow, it will hit the break point each time a new region
is launched. Then you can verify for each region whether only the PVMs
for this region are running.
- Check the logs of `ComputingUnitMaster` and see if `Region XX
successfully terminated.` is logged for each region.