feat(operator): treeplot operator (#3529)
## Overview  

This PR adds the tree plot operator to the current skeleton. This
operator provides a hierarchical overview of a bunch of data in the
workflow. It takes a column of parent-child relationships and
automatically generates a tree diagram.

## Input  

* **From Previous Workflow**: A table with columns where each entry is a
list or two-element tuple representing a `[parent, child]` edge (e.g.,
`['Root', 'BranchA']`).

* **From User**: The user need to specify which column in the table
needs to be organized in a tree plot manner. The user can select from a
dropdown menu, where all the columns are listed and users can click from
them.

## Output  

* A tree plot originated from up to down, representing all the
parent-child relationship in the given column. If two entries have the
same parent, then their parent nodes will be merged as one.

## Implementation  

* We mainly use igraph to generate plots and make the plot interactive.

## Features    

* Users can scroll up and down, left and right to adjust the viewpoint.
They can return to the defualt viewpoint by double clicking.
* Demo  


https://github.com/user-attachments/assets/aef4415a-9eae-4fcd-a8c6-36721a2370bd



## Next Steps  

* We plan to add the zoom-in/zoom-out feature, where users can more
intuitively use touchpad and two fingers to zoom in or zoom out the
plot.
* The canvas of the plot can be bigger.