chore(ci): Add Mock DB and JOOQ Generation to CI (#3689)
This PR introduces the following changes:

1. Adds a mock database for use in the CI environment.
2. Configures CI to generate JOOQ files before running tests.

We are planning to remove the JOOQ-generated files from the codebase.
However, the CI/CD pipeline currently relies on these files, and JOOQ
cannot regenerate them without a live database, which is not available
during CI/CD runs. This change addresses that limitation.

Additionally, the default configuration is updated to use the password
`postgres` instead of an empty string. This change is necessary because
the CI database does not permit empty passwords. It has no impact on
development or production environments, as developers and deployment
will override this default with their own credentials.

Closes issue #3688.