feat(frontend): introduce dropdown menu for layers (#3898)
Introduce a dropdown menu to toggle different information layers in the
workflow.

Currently, we only have a grid layer. However, additional layers—such as
workers, regions, and even physical operators—will be added in the
future. The dropdown will allow users to turn that information on or off
easily.


![Animation2](https://github.com/user-attachments/assets/5ede5a30-a506-4940-af5a-034b74d1a172)

fix #3897

---------

Signed-off-by: Xinyuan Lin <xinyual3@uci.edu>